chore: rename page
