+++ group/pages/group-detail-page.tsx
